requirements.txt
PyPDF2
tqdm
openai
soundfile
fastapi
uvicorn
python-multipart
numpy
requests
pathlib
google-genai
elevenlabs
anthropic
gradio
pydantic
python-docx
python-pptx
beautifulsoup4
trafilatura
